Make sure to select the correct task
A new service in the beginning of the pipeline (singleton-job) carries
the inputContainer to its outputContainer. This means that combined with
the collecting task, there are 2 tasks that fit this query. Only one
task, the first one which is often the singleton-job task, was returned.
This task was set to success instead of the collecting task. This meant
that the collecting task was never set to success and because of the
success of the singleton-job task, the collection task would be
respawned due to delta-notifier messages.